pyproject.toml
[tool.poetry]
name = "dairy-service"
version = "0.1.0"
description = "A dairy web service written using Flask/Python"
authors = ["Freedom Loisa <kitakayaloisa@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "dairy_service"}]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.3.2"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
flask-cors = "^4.0.0"
pytz = "^2023.3"
schedule = "^1.2.0"
ariadne = "^0.20.1"
pymongo = "^4.4.0"
bcrypt = "^4.0.1"
pyjwt = "^2.7.0"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
